✨ (App.jsx, NetworkDesign.jsx): refactor component imports to a new directory structure and remove unused components, enhancing organization and maintainability
