+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { MantineProvider } from '@mantine/core';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { Layout } from './components/Layout';
+import { Layout } from './components/layout/Layout';
 import { NetworkDesign } from './pages/NetworkDesign';
 
 /**
