+++ src/components/layout/NavbarMinimal.jsx
