+++ src/components/layout/NavbarMinimal.module.css
