+++ src/pages/NetworkDesign.jsx
@@ -2,13 +2,13 @@ import { useState, useEffect } from 'react';
 import { Container, TextInput, Button, Paper, Title, Grid, Text, Select, Group, Modal, Stack, Space } from '@mantine/core';
 import { useLocalStorage } from '@mantine/hooks';
 import { Netmask } from 'netmask';
-import { SubnetVisualization } from '../components/SubnetVisualization';
-import { ParentNetworkForm } from '../components/ParentNetworkForm';
-import { SubnetForm } from '../components/SubnetForm';
-import { DraggableSubnets } from '../components/DraggableSubnets';
-import { NetworkDiagram } from '../components/NetworkDiagram';
-import { NetworkDiagramSVGExport } from '../components/NetworkDiagramSVGExport';
-import { TerraformExportSection } from '../components/TerraformExportSection';
+import { SubnetVisualization } from '../components/tools/network-designer/SubnetVisualization';
+import { ParentNetworkForm } from '../components/tools/network-designer/ParentNetworkForm';
+import { SubnetForm } from '../components/tools/network-designer/SubnetForm';
+import { DraggableSubnets } from '../components/tools/network-designer/DraggableSubnets';
+import { NetworkDiagram } from '../components/tools/network-designer/NetworkDiagram';
+import { NetworkDiagramSVGExport } from '../components/tools/network-designer/NetworkDiagramSVGExport';
+import { TerraformExportSection } from '../components/tools/network-designer/TerraformExportSection';
 import { v4 as uuidv4 } from 'uuid';
 import { isValidIPv4, ipToLong, longToIp } from '../utils';
