+++ src/utils/regions/AzureRegions.js
